fix: changed the path of files one file above
