+++ money-manager/pom.xml
@@ -31,30 +31,33 @@
 	</properties>
 	<dependencies>
 
-		<!-- Dependency to persistence -->
+		<!-- (Java Persistence Api) Dependency used to facilitate the persistence in the Databank -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-data-jpa</artifactId>
 		</dependency>
 
+		<!-- Base to create the REST APIs in Spring -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
 		
-		<!-- PostgreSQL dependency -->
+		<!-- PostgreSQL (Databank) dependency -->
 		<dependency>
 			<groupId>org.postgresql</groupId>
 			<artifactId>postgresql</artifactId>
 			<scope>runtime</scope>
 		</dependency>
 
+		<!-- Facilitate the common tasks that sometimes in java may be too verbose -->
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
 			<optional>true</optional>
 		</dependency>
 
+		<!-- Unit Test dependency -->
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-test</artifactId>
