+++ money-manager/src/main/java/com/demo/MoneyManagerApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
