+++ money-manager/src/main/java/com/demo/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller;
+package com.demo.controller;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,8 +15,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.demo.model.User;
-import com.example.demo.worker.UserWorker;
+import com.demo.model.User;
+import com.demo.worker.UserWorker;
 
 @RestController
 @RequestMapping("user")
