+++ money-manager/src/main/java/com/demo/model/Card.java
@@ -1,6 +1,9 @@
-package com.example.demo.model;
+package com.demo.model;
+
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
@@ -9,23 +12,29 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.OneToMany;
 import lombok.Data;
+
 
 @Data
 @Entity
-public class Expense {
+public class Card {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	private Long id;
 
 	@ManyToOne
-	@JoinColumn(name = "card_id")
+	@JoinColumn(name = "wallet_id")
 	@JsonBackReference
-	private Card card;
+	private Wallet wallet;
 
-	private String name;
+	private String type;
+	private Float maxLimit;
+	private Float amount;
+	private String closeDay;
 
-	@OneToOne(mappedBy = "expense", cascade = CascadeType.ALL, orphanRemoval = true)
-	private Installment installment;
+	@OneToMany(mappedBy = "card", cascade = CascadeType.ALL, orphanRemoval = true)
+	@JsonManagedReference
+	private List<Expense> expenses;
+
 }
