+++ money-manager/src/main/java/com/demo/model/Installment.java
@@ -1,4 +1,4 @@
-package com.example.demo.model;
+package com.demo.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
