+++ money-manager/src/main/java/com/demo/model/Investment.java
@@ -1,4 +1,4 @@
-package com.example.demo.model;
+package com.demo.model;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
