+++ money-manager/src/test/java/com/demo/MoneyManagerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class MoneyManagerApplicationTests {
 
+
 	@Test
 	void contextLoads() {
+		
 	}
 
 }
